01_clear_data_generation_files script update: more robust. doesnt fail from missing files.

+++ data_generation_scripts/01_clear_data_generation_files.py
@@ -16,11 +16,19 @@ seperation_bar()
 print("Script 01: Clear Data Generation Files\n")
 
 try:
-    os.remove(data_generation_config_json_path)
-    print(f"[INFO] Data Generation Config JSON Deleted at '{data_generation_config_json_path}'")
-    
-    os.remove(generated_raw_match_data_json_path)
-    print(f"[INFO] Generated Raw Match Data JSON Deleted at '{generated_raw_match_data_json_path}'")
+    if os.path.exists(data_generation_config_json_path):
+        os.remove(data_generation_config_json_path)
+        print(f"[INFO] Data Generation Config JSON Deleted at '{data_generation_config_json_path}'")
+    else:
+        print(f"[INFO] File '{data_generation_config_json_path}' does not exist. Skipping.")
+
+    if os.path.exists(generated_raw_match_data_json_path):
+        os.remove(generated_raw_match_data_json_path)
+        print(f"[INFO] Generated Raw Match Data JSON Deleted at '{generated_raw_match_data_json_path}'")
+    else:
+        print(f"[INFO] File '{generated_raw_match_data_json_path}' does not exist. Skipping.")
+
+    print("\n[INFO] Script 01: Completed successfully.")
 
 except Exception as e:
     print(f"\n[ERROR] An error occurred: {e}")
